Server update for MyEclipse 10.
A key change is that inter project dependencies are now managed through maven in order to properly deploy.

MyEclipse deployment doesn't seem capable of jarring up dependent projects, but probably expects that to be handled through maven.

Also this uses the newer subclipse 1.8 line so workspaces are not compatible with pre svn 1.7 tools.